backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
python-docx
google-genai
groq
python-dotenv
rank-bm25
faiss-cpu
pymongo[srv]
bcrypt
pyjwt

